refactor(settings): split dialog coordinator helpers
Move settings dialog UI coordination, modal builders, event handlers, portable config support, and transfer actions into focused mixins while keeping SettingsDialogSimple as the public entry point.

Tested: settings dialog regression suite 92 passed; uv run ruff check src tests scripts; uv run pyright

Confidence: high
