+++ src/accessiweather/ui/dialogs/settings_dialog_constants.py
@@ -1,0 +1,7 @@
+"""Shared constants for settings dialog helpers."""
+
+API_KEYS_TRANSFER_NOTE = (
+    "API keys are not included here. API keys stay in this machine's secure keyring by "
+    "default. To transfer them, use 'Export API keys (encrypted)' and then "
+    "'Import API keys (encrypted)'."
+)
